Remove redundant assertion.
I guess we used to just have a g_assert() here to check that there was a
subdissector table to which to add the capture dissector, but we now
have a test for a null subdissector table pointer that attempts to
produce a more detailed and more useful error (and to fail only if we've
set the "crash on some otherwise-ignored errors" environment variable).

Remove the assertion, as we've already ensured that, at that point, the
subdissector table pointer is not null.

This should squelch a constArgument warning from cppcheck.