ansi_a: bsmap stats: populate table rows only once
Commit f4ac70818a updated the init function to use stat_tap_find_table and
create the statistics table only when it doesn't already exist.

However, if the table does already exist, its rows are still populated
again. If the table rows contain data that is allocated dynamically by
g_malloc or g_strdup, populating the table again (without freeing its rows
before) causes a memory leak.

If the init function is called and the statistics table already exists, all
we have to do is clear the data that was gathered when the tap parsed the
packets. This is what the stat_tap_reset_table_cb function does.

Fixes: f4ac70818a ("stat_tap_table_ui: create tables only once during init")