SOME/IP: Fix crash after loading a faulty uat config (BUGFIX)
When you load a correct wireshark config for SOME/IP responding
hashtable entries are created. If you load afterwards a new config
(empty or not) the old entries are updated.

However, Wireshark does not call the uat's post update callback,
when there was a bug in the config. This leads to an inconsistent
state, which may result in a crash on dissecting SOME/IP messages.

This patch adds code to the SOME/IP dissector to avoid inconsistent
state.

Fixes: #17227