WSDG: give more details on setting up a UNIX build environment.
List the minimum set of tools required.

We have scripts to do the setup work on a number of platforms.  Let the
user know about them.

Give instructions on using CMake; we're not using the traditional
autoconf stuff any more.

Give instructions on building the Developer's and User's Guides in the
UNIX section, and, in both that section *and* the equivalent Windows
section, give the name of the build target for building all guides.