gprs-ns: Mark NS-STATUS as in_error_pkt
Like many transport protocols, NS has the ability to include
the "erroneous message" when reporting errors to its peer in
NS-STATUS PDUs.

The current UX however is super annoying:  The BSSGP dissector
clears COL_INFO and hence if you look at the packet list in wireshark,
it looks like a valid higher-layer message is transmitted over NS,
and there is no mention that this is an error (NS-STATUS).

By simply setting in_error_pkt, the behavior changes: The erroneous
message is still dissected in the protocol details, but COL_INFO
remains what the NS decoder has to say: NS-STATUS with a decoded
cause information.