dvb-ci: pass ami payload to mime-encap dissector
Using the application mmi (ami) resource, a file of any type may be sent
from the module to the host.

The host receives both the file name and the binary payload. At the moment,
we parse the name and hand the payload to the png dissector if it ends with
.png

Instead of this manual approach, we should make use of all file types that
wireshark can dissect, i.e. all file types registered in the wtap_file
dissector table. The mime-encap dissector does just that, so we pass our
payload to this dissector.