arinc615a: Avoid calling g_str_has_suffix(NULL, ...).
Testing with tftpConversationError.pcapng attached to issue 10305 revealed this
warning on the console:

   GLib-CRITICAL **: 16:47:08.092: g_str_has_suffix: assertion 'str != NULL' failed

The cause is that the filename retrieved from the tftpinfo struct could
potentially be NULL when dissect_a615a_heur is called, for instance if the TFTP
RRQ/WRQ was not captured or not associated with the same conversation as the
DATA packet.

It's interesting that this condition arises from this capture
file... Perhaps the conversation tracking is amiss?  To be investigated.

Without knowing the filename, there appears to be no way to meaningfully
dissect the protocol beyond just dissecting just the file length and the
protocol version; For simplicity, I opted to maintain the present behavior and
have the heuristic test fail if the filename is not known.