Fix reassemble_test's additional debug mode.
It has bit-rotted.

Replace a bunch of compile-out "#if 0" with runtime "if (0)", and the "#ifdef
debug" with a static const, both of which should reduce the chance of this
bit-rotting again in future by ensuring that these code paths will at least be
compiled -- even if they are not actively used.  The default is kept at FALSE,
so the behavior is unchanged, and toggling it still requires recompilation.

fd->data no longer exists; The nearest equivalent uses (dangerous!) tvb_get_ptr
to dump the address of the first byte of the tvb data, in case that is somehow
useful... I'm guessing it exists for live debugging.

We no longer have access to the structures for the hash keys.  For the time
being, let's just #ifdef out the code that tries to print those keys.  Maybe we
should move the key structures to an epan/reassemble-int.h so that we can
access them from this test code again...?

And zap an extra comma which snuck into a table... Clearly this code hasn't
been compiled in a while.

Tested with debug=FALSE and debug=TRUE, under Valgrind as well as natively, on
a Linux host.