git: show expert info for unrecognized special packet 0003
In Git protocol, a pkt-line consists of a 4-hexdigit pkt-length,
followed by several bytes of pkt-data.  The pkt-length represents the
length of the entire pkt-line including the length field, so for an
ordinary pkt-line the length is always >= 4.  This allows the protocol
to use values less than 4 as special values --- for example, 0000 is a
so-called flush-pkt, representing the end of a command.

There's one particular pkt-length value that should never appear: 0003
is not >= 4 and is not a flush-pkt, delim-pkt, or response-end-pkt, so
it is not permitted in Git protocol.  Currently the dissector handles
this case by returning length 0 so it doesn't show up in wireshark as
Git protocol.  Better to treat it as Git protocol and add expert-info
describing what is wrong in case it shows up in a corrupt capture.

Part of #17093.  Based on a hint from Pascal Quantin at [1].

[1] https://gitlab.com/wireshark/wireshark/-/merge_requests/1946#note_515567051