MP2T: Detect stuffing bytes after SECTs when PUSI is 0
When the payload_unit_start_indicator is 0, a continued section may be
followed by stuffing bytes. Detect the stuffing bytes based on the SECT
length and handle them in the MPEG2 TS dissector in that case, rather than
handing them to the SECT dissector (which assumes that the incoming tvb
contains only one section and no additional data), similar to how stuffing
bytes are already handled when the PUSI is 1.