prism: clean up some things.
"CCK" really means "DSSS", whether it uses CCK or not, i.e.  it includes
both the legacy 802.11 DSSS PHY (1 Mb/s and 2 MB/s) and the 802.11b
HR/DSSS PHY (5.5 Mb/s and 11 Mb/s).  Change the text and the name of a
variable to match.

For OFDM, put the modulation type after the rate, as is done for the
text for other modulations.  Put "Rate:" before the rate for all
modulations as well.