Fix the captions for the buttons in the "export specified packets" dialog.
The "Selected packet" caption is wrong, as there could be more than one
packet selected.  Furthermore, in the non-Windows "export specified
packets" dialog, the caption is "Selected packets only", so go with the
full caption.

The "Marked packets" caption also lacks the "only" that it has in the
other dialog; add it as well.